pyproject.toml
[tool.poetry]
name = "strawberry-django-jwt2"
version = "0.2.4"
description = "Strawberry-graphql port of the graphene-django-jwt package"
license = "MIT"
authors = ["SOCGNA KOUYEM Childéric <socgnachilderic@proton.me>"]
maintainers = []
readme = "README.md"
classifiers = [
    'Development Status :: 3 - Alpha',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Framework :: Django',
    'Framework :: Django :: 3.2',
    'Framework :: Django :: 4.0',
    'Framework :: Django :: 5.0',
]
packages = [
    { include = "strawberry_django_jwt2" }
]

[tool.poetry.dependencies]
python = "^3.8"
Django = ">=4"
PyJWT = ">=1.7.1,<3.0"
strawberry-graphql = ">=0.121.1,<1.0"
strawberry-graphql-django = { version = ">=0.3", allow-prereleases = true }
django-admin-display = "^1.3.0"
packaging = ">=20.0,<30.0"
importlib-metadata = { version = ">=1.7,<5.0", python = "<=3.7" }
# For use in local tests and development, in future poetry release replace with dev-dependencies group.
uvicorn = { extras = ["standard"], version = "^0.30.6", optional = true }

[tool.poetry.group.dev.dependencies]
cryptography = ">=38.0"
coverage = { extras = ["toml"], version = "^6.4.1" }
pytest = "^7.1.3"
pytest-cov = ">=4.0"
pytest-django = "^4.0.0"
types-cryptography = "^3.3.15"
django-stubs = { extras = ["compatible-mypy"], version = "<1.12.0" }
django-mock-queries = "^2.1.7"
types-mock = "^4.0.0"
types-jwt = "^0.1.0"
# types-pkg-resources = "^0.2.2"
djangorestframework-stubs = "^1.4.0"
djangorestframework = "^3.15.2"
nox-poetry = "^1.0.1"
nox = "^2024.4.15"
mock = { version = "^4.0", python = "<3.8" }

[tool.poetry.extras]
dev = ["uvicorn"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "tests.example_app.settings"
python_files = "tests.py test_*.py *_tests.py"
filterwarnings = []

[tool.black]
line-length = 160
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
force_sort_within_sections = true

[tool.django-stubs]
django_settings_module = "strawberry_django_jwt2.settings"

[tool.coverage.run]
source = ["strawberry_django_jwt2"]
branch = true
omit = [".venv/**", "tests/migrations/**", "manage.py"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: nocover",
    "pragma:nocover",
    "if TYPE_CHECKING:",
    "@overload",
    "@abstractmethod",
    "@abc.abstractmethod",
    "assert_never",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/KundaPanda/strawberry-django-jwt/"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
